# docs/openapi/desktop-resources-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Devku Desktop Resource Distribution API",
    "version": "1.0.0",
    "description": "MCP/Skill publication and administration with private object storage. Resource storage shares Desktop update endpoint and credentials, using the explicitly configured private resource_bucket. Authorized users can obtain S3 presigned download URLs valid for 300 seconds; URLs can be reused until expiry.",
    "license": {
      "name": "LGPL-3.0",
      "url": "https://www.gnu.org/licenses/lgpl-3.0.html"
    }
  },
...
  "paths": {
    "/api/v1/admin/desktop/resources": {
      "post": {
        "operationId": "publishPublicDesktopResource",
        "tags": [
          "Publishing"
        ],
        "summary": "Upload one ZIP and publish it immediately",
        "description": "Raw ZIP <=64 MiB. Uploaded object first, then resource metadata commits atomically in PostgreSQL. New resources are active; publishing never changes an existing disabled status. Returns only the uploaded platform. Duplicate tuple or older version is 409. On ambiguous results verify through administrator key lookup and historical version SHA-256. Objects are uploaded to a private resource_bucket; no permanent download URL is returned or persisted.",
        "security": [
          {
            "adminBearer": []
          },
          {
            "adminApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/zip": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "description": "ZIP <=64 MiB; manifest.json <=128 KiB; <=1024 entries; <=128 MiB uncompressed."
        },
        "responses": {
          "200": {
            "description": "Published. data.artifacts contains only the uploaded platform.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "RESOURCE_PACKAGE_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "RESOURCE_VERSION_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Package/ingress body limit exceeded. Proxy responses may not be JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AdminErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Existing administrator authentication failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Existing administrator authorization failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "ADMIN_COMPLIANCE_ACK_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Other upstream or administrator middleware error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listAdminDesktopResources",
        "tags": [
          "Management"
        ],
        "summary": "List resources, including disabled resources",
        "security": [
          {
            "adminBearer": []
          },
          {
            "adminApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResourceListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Administrator authentication required"
          },
          "403": {
            "description": "Administrator access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Invalid query or status"
          },
          "423": {
            "description": "Administrator compliance acknowledgment required"
          },
          "default": {
            "description": "Service or proxy failure"
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number; maximum offset 10000.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Resources or versions per page.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "Resource type.",
            "schema": {
              "type": "string",
              "enum": [
                "",
                "mcp",
                "skill"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Resource status.",
            "schema": {
              "type": "string",
              "enum": [
                "",
                "active",
                "disabled"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Name/description search, max 200 UTF-8 bytes. % and _ are ILIKE wildcards.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": false,
            "description": "Exact resource key for publication reconciliation.",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9_-]{1,63}$"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/desktop/resources/validate": {
      "post": {
        "operationId": "validateDesktopResource",
        "tags": [
          "Management"
        ],
        "summary": "Validate a ZIP without publishing",
        "security": [
          {
            "adminBearer": []
          },
          {
            "adminApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Administrator authentication required"
          },
          "403": {
            "description": "Administrator access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Invalid query or status"
          },
          "423": {
            "description": "Administrator compliance acknowledgment required"
          },
          "default": {
            "description": "Service or proxy failure"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/zip": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "description": "ZIP <=64 MiB; manifest.json <=128 KiB; <=1024 entries; <=128 MiB uncompressed."
        }
      }
    },
...
    "/api/v1/admin/desktop/resources/{resource_id}/status": {
      "patch": {
        "operationId": "setDesktopResourceStatus",
        "tags": [
          "Management"
        ],
        "summary": "Disable or restore catalog and API access",
        "security": [
          {
            "adminBearer": []
          },
          {
            "adminApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Administrator authentication required"
          },
          "403": {
            "description": "Administrator access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Invalid query or status"
          },
          "423": {
            "description": "Administrator compliance acknowledgment required"
          },
          "default": {
            "description": "Service or proxy failure"
          }
        },
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^res_[0-9a-f]{32}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "disabled"
                    ]
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 2000,
                    "description": "Required and nonblank for disabled. Maximum 2000 UTF-8 bytes. Cleared on restore."
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "description": "Does not delete packages, uninstall local files or interrupt downloads already authorized. Blocks new download URL issuance; existing presigned URLs expire within 5 minutes. Publishing while disabled preserves disabled status."
      }
    },
    "/api/v1/admin/desktop/resources/{resource_id}/versions/{version}/artifacts/{platform}/download-url": {
      "post": {
        "operationId": "createAdminDesktopResourceDownloadURL",
        "tags": [
          "Management"
        ],
        "summary": "Authorize and issue a private download URL valid for 5 minutes",
        "description": "No request body. Resource must be active, including historical versions. Requires existing administrator or Desktop member authentication. The presigned URL can be reused until expiry; disabling a resource or revoking membership blocks further issuance but cannot revoke an already issued URL. Fetch the URL directly from S3 without User/Model Tokens. No URL is stored in resource metadata.",
        "security": [
          {
            "adminBearer": []
          },
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "description": "Opaque server resource ID.",
            "schema": {
              "type": "string",
              "pattern": "^res_[0-9a-f]{32}$"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "description": "Pinned version; historical artifacts remain downloadable.",
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]{0,5})\\.(0|[1-9][0-9]{0,5})\\.(0|[1-9][0-9]{0,5})$"
            }
          },
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "description": "Exact platform from artifact metadata.",
            "schema": {
              "type": "string",
              "enum": [
                "darwin-arm64",
                "darwin-x64",
                "windows-x64",
                "linux-x64",
                "any"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Short-lived download URL",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDownloadURLResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied, revoked membership or installation mismatch"
          },
          "404": {
            "description": "Resource/version/platform missing or resource disabled"
          },
          "422": {
            "description": "Invalid request or missing v2 installation ID"
          },
          "503": {
            "description": "Private resource storage is not configured or signing is unavailable"
          },
          "423": {
            "description": "Administrator compliance acknowledgment required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "desktopBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Desktop member User Token, never Model Token."
      },
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Existing administrator access token."
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Existing administrator API key, subject to administrator middleware."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "required": [
              "code",
              "message",
              "request_id",
              "retryable",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ],
        "examples": [
          {
            "error": {
              "code": "RESOURCE_NOT_FOUND",
              "message": "resource not found",
              "request_id": "",
              "retryable": false,
              "details": {}
            }
          }
        ]
      },
...
      "ValidationResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "const": 0
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Validation"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Sub2API deployment"
    }
  ]
}